refactor(svelte2tsx): 属性列の組み立てを attributes/mod.rs へ切り出す
build_attributes_string* / build_attribute_segments / build_component_props*
を純粋移動。祖先モジュールの私有項目は子孫から見えるため、PR で足していた
不要な pub(super) も併せて外した。出力・ロジックの変更なし。
